% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqSpikeFree.R
\name{ChIPseqSpikeFree}
\alias{ChIPseqSpikeFree}
\title{wrapper function - perform ChIP-seq spike-free normalization in one step.}
\usage{
ChIPseqSpikeFree(bamFiles = bams, chromFile = chrom.size,
  metaFile = metaFile, prefix = prefix)
}
\arguments{
\item{bamFiles}{a vector of bam filenames.}

\item{chromFile}{chrom.size file. Given "hg19","mm10","mm9" or "hg38", will load chrom.size file from package folder.}

\item{metaFile}{a filename of metadata file. the file must have three columns: ID (bam filename without full path), ANTIBODY and GROUP}

\item{prefix}{prefix of output filename.}
}
\value{
A data.frame of the updated metaFile with scaling factor
}
\description{
This function wraps all steps.  
If you run ChIPseqSpikeFree() seperately for two batches, the scaling factors will be not comparable between two batches. 
The correct way is to combine bamFiles parameter and create a new metadata file to include all bam files. Then re-run ChIPseqSpikeFree().
}
\examples{
##1 first You need to generate a sample_meta.txt (tab-delimited txt file).
metaFile <- "your/path/sample_meta.txt"
meta <- ReadMeta(metaFile)
head(meta)
ID    ANTIBODY    GROUP
ChIPseq1.bam  H3K27me3    WT
ChIPseq2.bam  H3K27me3    K27M

##2. bam files 
bams <- c("ChIPseq1.bam","ChIPseq2.bam")
prefix <- "test"

##3. run ChIPseqSpikeFree pipeline  
ChIPseqSpikeFree(bamFiles=bams, chromFile="mm9",metaFile=metaFile,prefix="test")

}
