% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqSpikeInFree.R
\name{PlotDistr}
\alias{PlotDistr}
\title{This function generates CPMW distribtion curve and barplot using sacaling factor table.}
\usage{
PlotDistr(data, SF = "test_SF.txt", prefix = "test", xlimMaxCPMW = NULL)
}
\arguments{
\item{data}{a data.frame generated by function ParseReadCounts() or a file name of parsed matrix}

\item{prefix}{prefix of output filename to save the plots and scaling factor values.}

\item{xlimMaxCPMW}{NULL or a numeric value [ between 10 and 100 ] to define maximum CPMW in the distribtion plot (xlim).}

\item{metaFile}{a data.frame of metadata by ReadMeta(); or a filename of metadata file.}
}
\value{
None
}
\description{
This function generates CPMW distribtion curve and barplot using sacaling factor table.
}
\examples{
## 1. start from a parsedMatrix file

# parsedMatrixFile <- "your/path/test_parsedMatrix.txt"
# metaFile <- "your/path/sample_meta.txt"
# parsedDF <- read.table(parsedMatrixFile, sep="\t",header=TRUE,fill=TRUE,
# quote="",row.names=NULL ,check.names=F)

## use defalut setting 
# SF <- CalculateSF(data = parsedDF, metaFile = metaFile)  
# PlotDistr (data=parsedDF,SF=SF, prefix="your/path/test", xlimMaxCPMW=NULL)

## use custom setting for H3K9me3
# SF <- CalculateSF(data = parsedDF, metaFile = metaFile, 
#     maxLastTurn=0.97, cutoff_QC=1)  
# PlotDistr (data=parsedDF,SF=SF, prefix="your/path/test_manual_cutoff", xlimMaxCPMW=NULL)

## zoom out distribution curve
# PlotDistr (data=parsedDF,SF=SF, prefix="your/path/test_manual_cutoff_zoom", xlimMaxCPMW=50)

}
