% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqSpikeInFree.R
\name{BoxplotSF}
\alias{BoxplotSF}
\title{This function generates boxplot using sacaling factor table. It's been included in the last step of ChIPseqSpikeInFree().}
\usage{
BoxplotSF(input, prefix = "test")
}
\arguments{
\item{input}{a file/data.frame generated/returned by CalculateSF() or ChIPseqSpikeInFree().
It looks like metadata file but has extra columns SF and COLOR.}

\item{prefix}{prefix of output filename.}
}
\value{
A filename of generated boxplot
}
\description{
This function generates boxplot using sacaling factor table. It's been included in the last step of ChIPseqSpikeInFree().
}
\examples{
## 1. re-generate boxplot of ChIPseqSpikeInFree scaling factors
## After you run ChIPseqSipkeFree(), a sacaling factor table
## (for example, test_SF.txt) will be generated.

# BoxplotSF(input="test_SF.txt",prefix="test")
}
