% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqSpikeFree.R
\name{CalculateSF}
\alias{CalculateSF}
\title{calculate scaling factors and save results}
\usage{
CalculateSF(data = data, metaFile = meta, prefix = prefix,
  dataRange = c(0.1, 0.99))
}
\arguments{
\item{data}{a data.frame generated by function ParseReadCounts() or a file name of parsed matrix}

\item{metaFile}{a data.frame of metadata by ReadMeta(); or a filename of metadata file.}

\item{prefix}{prefix of output filename to save the scaling factor values.}

\item{dataRange}{set c(0.1 and 0.99) by default. The first 10 percent of noisy weakest reads were excluded and stop calculation when 99 percent of reads are used.}
}
\value{
A data.frame of the updated metaFile with scaling factors
}
\description{
This function allows you to plot curves, caculate SF(scaling factor) per antibody based on parsedMatrix. 
In addtion return a data.frame of updated metadata.
If you run ChIPseqSpikeFree() seperately for two batches, the scaling factors will be not comparable between two batches. 
The correct way is to combine bamFiles parameter and create a new metadata file to include all bam files. Then re-run ChIPseqSpikeFree().
}
\examples{
##1. start from a parsedMatrix file

parsedMatrixFile <- "your/path/test_parsedMatrix.txt"
metaFile <- "your/path/sample_meta.txt"
parsedDF <- read.table(parsedMatrixFile, sep="\\t",header=TRUE,fill=TRUE,
 quote="",row.names=NULL ,check.names=F)
res <- CalculateSF (data=parsedDF,metaFile=metaFile, prefix="your/path/test")

##2. start from a rawCount file

metaFile <- "your/path/sample_meta.txt"
parsedDF <- ParseReadCounts(data="your/path/test_rawCounts.txt", metaFile=metaFile, 
     prefix="your/path/test_parsedMatrix.txt")
res <- CalculateSF (data=parsedDF,metaFile=metaFile, 
     prefix="your/path/test")

}
