% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqSpikeFree.R
\name{CountRawReads}
\alias{CountRawReads}
\title{count raw reads for all bins}
\usage{
CountRawReads(bamFiles = bams, chromFile = input, prefix = prefix,
  singleEnd = TRUE)
}
\arguments{
\item{chromFile}{a chrom.size file. Given "hg19","mm10","mm9" or "hg38", will load chrom.size file from package folder. Otherwise, give a /your/path/chrom.size}

\item{prefix}{prefix of output file name}

\item{singleEnd}{To count paired-end reads, set argument singleEnd=FALSE}

\item{bamfiles}{a vector of bam filenames.}
}
\value{
a data.frame of raw counts for each bin
}
\description{
This function counts raw reads for each bin.
}
\examples{
##1.count reads using mm9 bams 
bams <- c("your/path/ChIPseq1.bam","your/path/ChIPseq2.bam")
rawCountDF <- CountRawReads(bamFiles=bams,chromFile="mm9",prefix="your/path/test",singleEnd=TRUE)

}
